backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
openai
pydantic
passlib[bcrypt]
python-multipart
pyjwt
httpx
